refactor scripts and tests
