+++ scripts/generate_user_context.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 import os
 
-REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+from genlib import REPO_ROOT, HEADER_COMMENT
 
 FRONTEND_SRC = os.path.join(REPO_ROOT, "frontend", "src", "shared")
 
 def generate_user_context() -> str:
-    content: list[str] = ["import { createContext } from 'react';", "import type { UserData } from './RpcModels';", ""]
+    content: list[str] = HEADER_COMMENT + ["import { createContext } from 'react';", "import type { UserData } from './RpcModels';", ""]
 
     # UserData interface is defined in RpcModels.tsx
 
@@ -32,9 +32,9 @@ def generate_user_context() -> str:
 
     return "\n".join(content)
 
-def main() -> None:
-  os.makedirs(FRONTEND_SRC, exist_ok=True)
-  out_path = os.path.join(FRONTEND_SRC, "UserContext.tsx")
+def main(output_dir: str = FRONTEND_SRC) -> None:
+  os.makedirs(output_dir, exist_ok=True)
+  out_path = os.path.join(output_dir, "UserContext.tsx")
   with open(out_path, "w") as f:
     f.write(generate_user_context())
   print(f"✅ Wrote UserContext to '{out_path}'")
