+++ scripts/genlib.py
@@ -1,33 +1,32 @@
-# from typing import Any, get_type_hints
-# from pydantic import BaseModel
-
-# PY_TO_TS = {
-#   str: 'string',
-#   int: 'number',
-#   float: 'number',
-#   bool: 'boolean',
-#   list: "array",
-#   dict: "object",
-#   type(None): "null",
-# }
-
-# def field_to_ts(name: str, annotation: Any) -> str:
-#   """Convert a single Pydantic field to TypeScript."""
-#   ts_type = PY_TO_TS.get(annotation, 'any')
-#   return f"  {name}: {ts_type};"
-
-# def model_to_ts(cls: type[BaseModel]) -> str:
-#   fields = []
-#   hints = get_type_hints(cls, include_extras=True)
-#   for name, typ in hints.items():
-#     fields.append(field_to_ts(name, typ))
-#   body = "\n".join(fields)
-#   return f"export interface {cls.__name__} {{\n{body}\n}}\n"
-
-# genlib.py
+"""Shared helpers for code generation scripts."""
 
 from typing import Any, Union, get_origin, get_args
 from pydantic import BaseModel
+import os, importlib.util
+
+# Root of the repository relative to this file
+REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+
+# Standard header included at the top of generated files
+HEADER_COMMENT = [
+  "// ================================================",
+  "// WARNING: This file is automatically generated.",
+  "// Do not modify it by hand. Your changes will be",
+  "// overwritten the next time the generator runs.",
+  "// ================================================",
+  "",
+]
+
+def camel_case(name: str) -> str:
+  """Convert snake_case names to CamelCase."""
+  return "".join(part.capitalize() for part in name.split("_"))
+
+def load_module(path: str):
+  """Import a Python module from *path* and return it."""
+  spec = importlib.util.spec_from_file_location("mod", path)
+  module = importlib.util.module_from_spec(spec)
+  spec.loader.exec_module(module)  # type: ignore[attr-defined]
+  return module
 
 # Map bare Python types to TS primitives (won't be used for generic types)
 PY_TO_TS = {
